[Update]: refactoring MapController...
