+++ GurunaviAPI/Intro/SplashViewController.swift
@@ -11,7 +11,6 @@ class SplashViewController: UIViewController {
 
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
-        print("はろ")
         chooseShouldLaunchViewController()
     }
